+++ src/App.js
@@ -1,5 +1,5 @@
-const InputView = require('./InputView');
-const OutputView = require('./OutputView');
+const InputView = require('./views/InputView');
+const OutputView = require('./views/OutputView');
 const BridgeGame = require('./BridgeGame');
 const { COMMAND } = require('./constants/bridge');
