+++ src/views/InputView.js
@@ -1,6 +1,6 @@
-const Io = require('./utils/io');
-const { INPUT } = require('./constants/views');
-const { Validator } = require('./utils/validator');
+const Io = require('../utils/io');
+const { INPUT } = require('../constants/views');
+const { Validator } = require('../utils/validator');
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
